pop-it 0.1

+++ controllers/card.js
@@ -1,12 +1,12 @@
-import Card from "../models/card.js";
+import Card from '../models/card.js';
 
-const notFoundError = new Error("Запрашиваемая карточка не найдена");
-notFoundError.name = "NotFound";
-notFoundError.message = "Запрашиваемая карточка не найдена";
+const notFoundError = new Error('Запрашиваемая карточка не найдена');
+notFoundError.name = 'NotFound';
+notFoundError.message = 'Запрашиваемая карточка не найдена';
 
 export const getCards = async (req, res) => {
   try {
-    const cards = await Card.find({}).populate("owner");
+    const cards = await Card.find({}).populate('owner');
     res.send(cards);
   } catch (err) {
     res.status(500).send(err);
@@ -20,8 +20,8 @@ export const createCard = async (req, res) => {
     const newCard = await Card.create({ name, link, owner: _id });
     res.send(newCard);
   } catch (err) {
-    if (err.name === "ValidationError") {
-      res.status(400).send({ message: "Переданы некорректные данные" });
+    if (err.name === 'ValidationError') {
+      res.status(400).send({ message: 'Переданы некорректные данные' });
       return;
     }
 
@@ -32,21 +32,29 @@ export const createCard = async (req, res) => {
 export const deleteCard = async (req, res) => {
   try {
     const { cardId } = req.params;
+    const cardToDelete = await Card.findById({ _id: cardId });
+    if (!cardToDelete) {
+      throw notFoundError;
+    }
+    if (cardToDelete.owner !== req.user._id) {
+      throw new Error('Не достаточно прав');
+    }
     const deleteStatus = await Card.deleteOne({ _id: cardId });
     if (deleteStatus.deletedCount === 0) {
       throw notFoundError;
     }
     res.send(deleteStatus);
   } catch (err) {
-    if (err.name === "NotFound") {
+    if (err.name === 'NotFound') {
       res.status(404).send({ message: err.message });
       return;
     }
-    if (err.name === "CastError") {
-      res.status(400).send({ message: "Переданы некорректные данные" });
+    if (err.name === 'CastError') {
+      res.status(400).send({ message: 'Переданы некорректные данные' });
       return;
     }
-    res.status(500).send(err);
+    console.log(err)
+    res.status(500).send({ message: err.message });
   }
 };
 
@@ -59,19 +67,19 @@ export const addLike = async (req, res) => {
       {
         $addToSet: { likes: _id },
       },
-      { new: true }
-    ).populate(["owner", "likes"]);
+      { new: true },
+    ).populate(['owner', 'likes']);
     if (updatedCard === null) {
       throw notFoundError;
     }
     res.send(updatedCard);
   } catch (err) {
-    if (err.name === "NotFound") {
+    if (err.name === 'NotFound') {
       res.status(404).send({ message: err.message });
       return;
     }
-    if (err.name === "CastError") {
-      res.status(400).send({ message: "Переданы некорректные данные" });
+    if (err.name === 'CastError') {
+      res.status(400).send({ message: 'Переданы некорректные данные' });
       return;
     }
     res.status(500).send(err);
@@ -87,19 +95,19 @@ export const removeLike = async (req, res) => {
       {
         $pull: { likes: _id },
       },
-      { new: true }
-    ).populate(["owner", "likes"]);
+      { new: true },
+    ).populate(['owner', 'likes']);
     if (updatedCard === null) {
       throw notFoundError;
     }
     res.send(updatedCard);
   } catch (err) {
-    if (err.name === "NotFound") {
+    if (err.name === 'NotFound') {
       res.status(404).send({ message: err.message });
       return;
     }
-    if (err.name === "CastError") {
-      res.status(400).send({ message: "Переданы некорректные данные" });
+    if (err.name === 'CastError') {
+      res.status(400).send({ message: 'Переданы некорректные данные' });
       return;
     }
     res.status(500).send(err);
